This project is on the multiple objects detection in images with YOLO v3 model. Here a pre-trained YOLO v3 model (trained with huge COCO dataset) has been used to detect various object in images. It not only detects objects in a image but also it creates a bounding box around a particular object and also writes the name of the class that object belongs to (cat/dog/horse/person etc.). It also writes the score (probability) with which that object was detected.
This project is on the multiple objects detection in images with YOLO v3 model. Here a pre-trained YOLO v3 model (trained with huge COCO dataset) has been used to detect various object in images. It not only detects objects in a image but also it creates a bounding box around a particular object and also writes the name of the class that object belongs to (cat/dog/horse/person etc.). It also writes the score (probability) with which that object was detected. YOLO v3 model is based on deep learning model (CNN).
Here, we downloaded 3 different images (dog-bicycle-person.jpg, person-car-horse.jpg, persons-car-dog.jpg) from https://unsplash.com/ for objects detection with YOLO v3.
Original images can be found in \images\test and result images (after object detection) can be found in \images\res. Results are quite satisfactory. It could detect all the relevant objects in the images successfully.